# src/swagger/vj-ota-api.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1",
    "title": "VietJet API GATEWAY",
    "description": "VietJet API GATEWAY",
    "termsOfService": "None",
    "contact": {
      "name": "VietJet API GATEWAY"
    },
    "license": {
      "name": "All right reserved"
    }
  },
  "servers": [
    {
      "url": "https://34.144.235.95.nip.io/flight"
    }
  ],
  "paths": {
...
    "/reservations": {
      "get": {
        "tags": ["Reservations"],
        "summary": "Search reservations for the provided criteria. Use the optional earliest/latest groups of parameters to return options for a date range up to a maximum of 31 days. At least one parameter must be provided not including partialMatch.",
        "operationId": "ApiReservationsGet",
        "parameters": [
          {
            "name": "reservationNumber",
            "in": "query",
            "description": "Optional number which identifies a reservation.",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "reservationLocator",
            "in": "query",
            "description": "Optional locator which identifies or is associated to a reservation.",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "earliestBooking",
            "in": "query",
            "description": "Optional earliest date of booking.",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "latestBooking",
            "in": "query",
            "description": "Optional latest date of booking.",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "departureAirport",
            "in": "query",
            "description": "Optional departure airport code.",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "earliestDeparture",
            "in": "query",
            "description": "Optional earliest date of departure.",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "latestDeparture",
            "in": "query",
            "description": "Optional latest date of departure.",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "passengerLastName",
            "in": "query",
            "description": "Optional last name of a reserved passenger.",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "passengerFirstName",
            "in": "query",
            "description": "Optional first name of a reserved passenger.",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "partialMatch",
            "in": "query",
            "description": "Optional partial matching of search parameters.",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservationModel"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Invalid Credentials": {
                      "type": "string",
                      "example": "Occurs when the api username and/or password are not valid for the endpoint you are trying to reach"
                    },
                    "Permission Denied": {
                      "type": "string",
                      "example": "Occurs when you do not have permission to perform the target action"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": ["Reservations"],
        "summary": "Add the provided reservation or split an existing reservation.",
        "operationId": "ApiReservationsPost",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "1000": {
                      "type": "string",
                      "example": "Credit Card processing issue."
                    },
                    "Deserialization Failed": {
                      "type": "string",
                      "example": "Happens when Request is not properly formatted (i.e. missing or misspelled elements"
                    },
                    "An error occurred when verifying security for the message": {
                      "type": "string",
                      "example": "Occurs when request is missing required element"
                    },
                    "ContractFilter mismatch at the EndpointDispatcher": {
                      "type": "string",
                      "example": "Can manifest because of a variety of different reasons, usually occurs when the endpoint URL is invalid"
                    },
                    "The booking request failed. The reservation was not created successfully": {
                      "type": "string",
                      "example": "Occurs when trying to book a reservation with credit card payment type if processing fees are applicable but not provided in the request"
                    },
                    "-91": {
                      "type": "string",
                      "example": "Object variable or With block variable not set - Occurs for a variety of different reasons, most common being that a required field is passed a null value."
                    },
                    "Company not found or invalid": {
                      "type": "string",
                      "example": "Occurs when searching for a company that does not exist in the database. Can also occur during booking with the Account payment type providing invalid company information"
                    },
                    "Cannot exceed more then + or 7 day": {
                      "type": "string",
                      "example": "Occurs when trying to get travel options passing a value greater than 7 for DaysBefore or DaysAfter"
                    },
                    "Failed to find reservation number: ReservationNumber": {
                      "type": "string",
                      "example": "Occurs when searching for a reservation that does not exist in the database (ReservationNumber will be the invalid reservation number that was passed into the request"
                    },
                    "Failed to add UserProfile to the database": {
                      "type": "string",
                      "example": "Occurs when attempting to add a user profile whose Profile Name already exists in the database (must be unique)"
                    },
                    "Failed to update user profile: ConcurrencyError": {
                      "type": "string",
                      "example": "Occurs when attempting to update a user profile that is being modified by another use"
                    },
                    "There was an error during login. Please try again": {
                      "type": "string",
                      "example": "Occurs during session login when the server if unreachable"
                    },
                    "-7001": {
                      "type": "string",
                      "example": "Fare has sold out during reservation attempt."
                    },
                    "-7002,-6955": {
                      "type": "string",
                      "example": "Seat availability has sold out during reservation attempt."
                    },
                    "-9987": {
                      "type": "string",
                      "example": "Infant seat availability has sold out during reservation attempt."
                    },
                    "-2147220500": {
                      "type": "string",
                      "example": "Purchase order is required and not provided."
                    },
                    "-101, -9799": {
                      "type": "string",
                      "example": "Company does not have enough credit for the purchase."
                    },
                    "-111": {
                      "type": "string",
                      "example": "Generalized booking error. This is a catch all error and does not specify any one cause."
                    },
                    "-10001": {
                      "type": "string",
                      "example": "Error getting user agency / Error getting reservation currency / currency conversion failure / Error getting company / Error getting credit card Type / Error getting payment method / Voucher payment type code is not equal to VO / Error identifying with fare class"
                    },
                    "-7777777": {
                      "type": "string",
                      "example": "Permission denied. The user is missing a required permission to complete the booking."
                    },
                    "-7006": {
                      "type": "string",
                      "example": "Error getting fare / No fares found for given booking code"
                    },
                    "-7009": {
                      "type": "string",
                      "example": "Unable to convert fares to desired currency."
                    },
                    "-6998": {
                      "type": "string",
                      "example": "The reservation balance is greater than the given payment."
                    },
                    "-6966": {
                      "type": "string",
                      "example": "Voucher attempt failed validation."
                    },
                    "-6955": {
                      "type": "string",
                      "example": "Availability check failed, either no valid fare was found or one or more of the seats are no longer available."
                    },
                    "-6956": {
                      "type": "string",
                      "example": "Error occurred adding reservation to the database."
                    },
                    "-2147217900": {
                      "type": "string",
                      "example": "Database foreign key constraint issue."
                    },
                    "-200001": {
                      "type": "string",
                      "example": "Failed Seat assignment."
                    },
                    "-120": {
                      "type": "string",
                      "example": "Error creating passenger profile."
                    },
                    "-65987": {
                      "type": "string",
                      "example": "This error generally manifests when the processing fees for the credit card payment are either not included or incorrect."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Invalid Credentials": {
                      "type": "string",
                      "example": "Occurs when the api username and/or password are not valid for the endpoint you are trying to reach"
                    },
                    "Permission Denied": {
                      "type": "string",
                      "example": "Occurs when you do not have permission to perform the target action"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/reservations/{key}": {
      "get": {
        "tags": ["Reservations"],
        "summary": "Retrieve the reservation for the provided key.",
        "operationId": "ApiReservationsByKeyGet",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "description": "The access key for a reservation.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Invalid Credentials": {
                      "type": "string",
                      "example": "Occurs when the api username and/or password are not valid for the endpoint you are trying to reach"
                    },
                    "Permission Denied": {
                      "type": "string",
                      "example": "Occurs when you do not have permission to perform the target action"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "api-key": {
        "type": "apiKey",
        "name": "apikey",
        "in": "header"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "This is method for login user"
      }
    },
    "schemas": {
...
      "CurrencyModel": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "baseCurrency": {
            "type": "boolean"
          },
          "currentExchangeRate": {
            "format": "double",
            "type": "number"
          },
          "format": {
            "type": "string"
          }
        }
      },
...
      "CompanyModel": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "identifier": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "account": {
            "$ref": "#/components/schemas/CompanyAccountModel"
          }
        }
      },
...
      "ReservationPassengerModel": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "index": {
            "format": "int32",
            "type": "integer"
          },
          "reference": {
            "type": "string"
          },
          "reservationOwner": {
            "type": "boolean"
          },
          "reservationStatus": {
            "$ref": "#/components/schemas/ReservationStatusModel"
          },
          "fareApplicability": {
            "$ref": "#/components/schemas/ChargePassengerApplicabilityModel"
          },
          "reservationProfile": {
            "$ref": "#/components/schemas/ReservationPassengerProfileModel"
          },
          "frequentFlyer": {
            "type": "object"
          },
          "advancePassengerInformation": {
            "type": "object"
          },
          "passengerServiceRequests": {
            "uniqueItems": false,
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "passengerCostCenter": {
            "type": "object"
          },
          "infants": {
            "uniqueItems": false,
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "weight": {
            "type": "object"
          },
          "notes": {
            "type": "string"
          },
          "journeys": {
            "uniqueItems": false,
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "paymentTransactions": {
            "uniqueItems": false,
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
...
      "ReservationModel": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "number": {
            "format": "int32",
            "type": "integer"
          },
          "locator": {
            "type": "string"
          },
          "bookingInformation": {
            "$ref": "#/components/schemas/BookingInfoModel"
          },
          "reservationSummary": {
            "type": "object"
          },
          "passengers": {
            "uniqueItems": false,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReservationPassengerModel"
            }
          },
          "journeys": {
            "uniqueItems": false,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReservationJourneyModel"
            }
          },
          "insurancePolicies": {
            "uniqueItems": false,
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "ancillaryPurchases": {
            "uniqueItems": false,
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "seatSelections": {
            "uniqueItems": false,
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "passengerLegDetails": {
            "uniqueItems": false,
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "charges": {
            "uniqueItems": false,
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "paymentTransactions": {
            "uniqueItems": false,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReservationPaymentTransactionModel"
            }
          },
          "eTickets": {
            "uniqueItems": false,
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "BookingInfoModel": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "sessionIdentifier": {
            "type": "string"
          },
          "distributionChannel": {
            "type": "object"
          },
          "externalLocators": {
            "uniqueItems": false,
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "agency": {
            "$ref": "#/components/schemas/AgencyModel"
          },
          "company": {
            "$ref": "#/components/schemas/CompanyModel"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyModel"
          },
          "contactInformation": {
            "$ref": "#/components/schemas/BookingContactInfoModel"
          },
          "creation": {
            "type": "object"
          },
          "hold": {
            "$ref": "#/components/schemas/BookingHoldModel"
          },
          "cancellation": {
            "type": "object"
          },
          "bookingType": {
            "type": "object"
          },
          "frequentFlyer": {
            "type": "object"
          },
          "notes": {
            "type": "string"
          },
          "paymentTransactions": {
            "uniqueItems": false,
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
...
      "BookingHoldModel": {
        "type": "object",
        "properties": {
          "overridden": {
            "type": "boolean"
          },
          "expiryTime": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "api-key": []
    }
  ]
}